add features

+++ Back-end/src/main/java/com/example/demo/repository/PasswordResetTokenRepository.java
@@ -4,7 +4,7 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import com.example.demo.entity.PasswordResetToken;
 import com.example.demo.entity.User;
 
-public interface PasswordResetTokenRepository extends JpaRepository<PasswordResetToken, Integer> {
+public interface PasswordResetTokenRepository extends JpaRepository<PasswordResetToken, Long> {
     PasswordResetToken findByToken(String token);
     PasswordResetToken findByUser(User user);
 }
